// src/interfaces/reference-deposits.interface.ts
export interface ReferenceDepositsInterface {
  reference_deposit_id: string;
  date_of_deposit: string;
  bank_name: string;
  amount_deposited: string;
  account_number: string;
  receipt_id: string;
  description: string;
  depositor: string;
  imei: string;
  app_version: string;
  time_of_deposit: string;
  status: number;
  hub_id: string;
  rejection_comment: string;
  rejection_date: string;
}